feat(GCPVpcPeering) Delete VPC Peering

+++ pkg/kcp/provider/gcp/vpcpeering/deleteVpcPeering.go
@@ -10,17 +10,22 @@ func deleteVpcPeering(ctx context.Context, st composed.State) (error, context.Co
 	obj := state.ObjAsVpcPeering()
 	logger := composed.LoggerFromCtx(ctx)
 
-	logger.Info("Deleting GCP VPC Peering " + obj.Spec.VpcPeering.Gcp.PeeringName)
+	if state.kymaVpcPeering == nil {
+		logger.Info("VPC Peering is not loaded")
+		return nil, nil
+	}
+
+	logger.Info("Deleting GCP VPC Peering " + obj.Spec.VpcPeering.Gcp.RemotePeeringName)
 
 	_, err := state.client.DeleteVpcPeering(
 		ctx,
-		&obj.Spec.VpcPeering.Gcp.PeeringName,
-		&state.Scope().Spec.Scope.Gcp.Project,
-		&state.Scope().Spec.Scope.Gcp.VpcNetwork,
+		obj.Spec.VpcPeering.Gcp.RemotePeeringName,
+		state.Scope().Spec.Scope.Gcp.Project,
+		state.Scope().Spec.Scope.Gcp.VpcNetwork,
 	)
 
 	if err != nil {
-		return err, ctx
+		return err, nil
 	}
 
 	return nil, nil
